scratch_tirg/requirements.txt
numpy
pillow
torch
torchvision
tqdm
tensorboardX

